+++ CHANGELOG.md
@@ -26,6 +26,8 @@
     - Question `display_type` and AnswerTemplate `data_type` and `allow_multiple` are no longer used to determine how to display questions on surveys
     - AnswerTemplates now allow a `template_name` to be specified of the following types:
       - "date", "radio", "checkbox", "string", "height", "number"
+  - Encounters can be added to surveys in online builder
+    - An encounter specifies when the survey is launched, in days after sign up
 - **Admin Changes**
   - Administrators can now unlock surveys for users from the user show page
 - **Gem Changes**